Added permission for updating or deleting threads

+++ app/Controllers/Http/ThreadsController.ts
@@ -52,9 +52,17 @@ export default class ThreadsController {
     }
   }
 
-  public async update({ params, request, response }: HttpContextContract) {
+  public async update({ params, auth, request, response }: HttpContextContract) {
     try {
+      const user = await auth.user
       const thread = await Thread.findOrFail(params.id)
+
+      if (user?.id !== thread.userId) {
+        return response.status(401).json({
+          message: 'You are not authorized to do this action'
+        })
+      }
+
       const validateData = await request.validate(ThreadValidator)
 
       await thread.merge(validateData).save()
@@ -72,9 +80,17 @@ export default class ThreadsController {
     }
   }
 
-  public async destroy({ params, response }: HttpContextContract) {
+  public async destroy({ params, auth, response }: HttpContextContract) {
     try {
+      const user = await auth.user
       const thread = await Thread.findOrFail(params.id)
+
+      if (user?.id !== thread.userId) {
+        return response.status(401).json({
+          message: 'You are not authorized to do this action'
+        })
+      }
+
       await thread.delete()
       return response.status(200).json({
         message: 'Thread deleted successfully',
